12bit turntable

+++ firmware/src/turntable.h
@@ -10,7 +10,7 @@
 #include <stdbool.h>
 
 void turntable_init();
-uint8_t turntable_read();
+uint16_t turntable_read();
 uint16_t turntable_raw();
 void turntable_update();
